# Optimize writing history file