- If we are not doing search in history but only
   add the last backup, there is no need to load
   tho whole file in the memory just to add one entry.
   Also there is no need to re-sort the history every
   time we write it as the latest entry always goes
   in the beginning of the file.

Tested on ~100 Mb gpbackup_history file
 `gpbackup —dbname postgres` command
Time between Post-data metadata backup complete and Email report

Timing from the end of backupPostdata to the beginning of emailReport
400Mb, master
1) 2 min 14.642 sec
2) 2 min 21.368 sec
3) 2 min 28.188 sec

400Mb, fix branch
1) Time: 2.033 sec
2) Time: 2.090 sec
3) Time: 2.157 sec

Timing for writing backup history only
without fix
1050mb 10m13.70491183s
with fix
1050mb 1.550556176s

Authored-by: Kate Dontsova <edontsova@pivotal.io>